hyperflexsizer-NewVersions/sizer/requirements.txt
Django==2.2.9
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
jsonfield==2.0.2
mailer==0.8.1
mysqlclient==1.4.6
packaging==16.8
Pillow==7.0.0
PuLP==1.6.8
reportlab==3.0
requests==2.22.0
simplejson==3.17.0
six==1.14.0
python-pptx==0.6.11
python-ldap==3.2.0
XlsxWriter==0.9.4
wget==3.2
elasticsearch==6.3.1
xlrd==1.2.0
shortuuid==1.0.1
